Modified HeadlessController: instead of exiting system, returns instead due to inability to test. Abstracted InteractiveControllerTest and HeadlessControllerTest due to egregious code duplication. Incorporated mock view into tests. Might scrap Interaction due to confusing line breaks with actual vs expected outputs.
