+++ Calendar/test/controller/tests/Interaction.java
@@ -1,8 +1,10 @@
+package controller.tests;
+
 /**
  * An interaction with the user consists of some input to send the program
  * and some output to expect.  We represent it as an object that takes in two
  * StringBuilders and produces the intended effects on them
  */
-interface Interaction {
+public interface Interaction {
   void apply(StringBuilder in, StringBuilder out);
 }
